requirements.txt
python-dotenv
azure-identity
azure-ai-agents
agent-framework